test: AbstractTestContainers configured (#14)
