+++ src/test/README.md
@@ -1,12 +1,6 @@
 e2e тесты пишутся по шаблону.  
-Если нужно поднять только контейнер postgres, наследуемся от AbstractPostgresTestContainers. Используем всегда, если не
-используется minio.  
-Если нужно поднять контейнер postgres и minio, наследуемся от AbstractPostgresMinioTestContainers. Используем, если
-нужно протестировать minio.  
-В этих классах вся конфигурация контейнеров необходимая для тестов.  
-
-На данный момент все тесты можно писать в IntegrationTest, если не нужно тестировать minio, и в IntegrationMinioTest, если тестируете minio.  
-Это позволит поднимать контейнер один раз на все тесты и сильно сэкономить всемя на прогонку тестов.
+Наследуемся от AbstractTestContainers. В этом классе вся конфигурация контейнеров необходимая для тестов.  
+Это позволит поднимать контейнеры postgres и minio один раз на все тесты и сильно сэкономить всемя на прогонку тестов.
 
 Для тестирования minio создаем всегда один bucket с именем "test-bucket", для этого существует константа MINIO_BUCKET.  
 После каждого теста все данные в таблицах postgres очищаются и удаляется bucket minio.
