+++ src/test/java/org/itmo/eventApp/main/controller/AbstractTestContainers.java
@@ -1,18 +1,19 @@
 package org.itmo.eventApp.main.controller;
 
 import io.minio.MinioClient;
+import io.minio.RemoveBucketArgs;
 import org.itmo.eventapp.main.Main;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
+import org.testcontainers.containers.MinIOContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 
 import javax.sql.DataSource;
@@ -22,12 +23,9 @@ import java.io.IOException;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-@ActiveProfiles("test-postgres")
+@ActiveProfiles("test")
 @ContextConfiguration(classes = Main.class)
-public abstract class AbstractPostgresTestContainers {
-    @MockBean
-    MinioClient minioClient;
-
+public abstract class AbstractTestContainers {
     @Autowired
     private DataSource dataSource;
 
@@ -36,11 +34,28 @@ public abstract class AbstractPostgresTestContainers {
 
     private final static String POSTGRES_VERSION = "postgres:16.0";
 
+    /**
+     * In tests create bucket only with this name
+     * After each test bucket with this name will be deleted
+     */
+    public final static String MINIO_BUCKET = "test-bucket";
+
     private static final PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>(POSTGRES_VERSION)
             .withUsername("test_user")
             .withPassword("test_password")
             .withDatabaseName("test_db")
             .withReuse(true);
+
+    private static final MinIOContainer minioContainer = new MinIOContainer("minio/minio")
+            .withUserName("test_minio_admin")
+            .withPassword("test_minio_admin")
+            .withReuse(true);
+
+    MinioClient minioClient = MinioClient
+            .builder()
+            .endpoint(minioContainer.getS3URL())
+            .credentials(minioContainer.getUserName(), minioContainer.getPassword())
+            .build();
 
     @BeforeAll
     public static void startPostgres() {
@@ -50,9 +65,22 @@ public abstract class AbstractPostgresTestContainers {
         System.setProperty("DB_PASSWORD", postgreSQLContainer.getPassword());
     }
 
+    @BeforeAll
+    public static void startMinio() {
+        minioContainer.start();
+        System.setProperty("MINIO_URL", minioContainer.getS3URL());
+        System.setProperty("MINIO_ACCESS_KEY", minioContainer.getUserName());
+        System.setProperty("MINIO_SECRET_KEY", minioContainer.getPassword());
+    }
+
     @BeforeEach
-    public void cleanUp() {
-        executeSqlScript("/sql/cleanTables.sql");
+    public void cleanUp() throws Exception {
+        try {
+            executeSqlScript("/sql/cleanTables.sql");
+            minioClient.removeBucket(RemoveBucketArgs.builder().bucket(MINIO_BUCKET).build());
+        } catch (Exception ignored) {
+
+        }
     }
 
     /**
