--- src/test/java/org/itmo/eventApp/main/controller/IntegrationMinioTest.java
@@ -1,5 +1,0 @@
-package org.itmo.eventApp.main.controller;
-
-public class IntegrationMinioTest extends AbstractPostgresMinioTestContainers {
-
-}
