--- src/test/java/org/itmo/eventApp/main/controller/IntegrationTest.java
@@ -1,5 +1,0 @@
-package org.itmo.eventApp.main.controller;
-
-public class IntegrationTest extends AbstractPostgresTestContainers {
-
-}
