--- src/test/resources/application-test-postgres.yml
@@ -1,23 +1,0 @@
-spring.application.name: event-app
-
-spring:
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: ${DB_URL}
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-  jpa:
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-        format_sql: true
-    show_sql: true
-  flyway:
-    enabled: true
-    locations:
-      - 'classpath:db/test-migration'
-
-minio:
-  url:
-  accessKey:
-  secretKey:
